+++ tests/test_prompt_template.py
@@ -3,7 +3,7 @@ import datetime
 import json
 import copy
 import pytest
-from memor import PromptTemplate, MemorValidationError
+from memor import PromptTemplate, MemorValidationError, MemorRenderError
 
 TEST_CASE_NAME = "PromptTemplate tests"
 
@@ -383,3 +383,27 @@ def test_size():
     template.save("template_test3.json")
     assert os.path.getsize("template_test3.json") == template.size
     assert template.size == template.get_size()
+
+
+def test_render1():
+    template = PromptTemplate(
+        content="Act as a {language} developer and respond to this question:\n{prompt_message}",
+        custom_map={"language": "Python"})
+    assert template.render({"prompt_message": "Mock Question"}
+                           ) == "Act as a Python developer and respond to this question:\nMock Question"
+
+
+def test_render2():
+    template = PromptTemplate(
+        content="Act as a {language} developer and respond to this question:\n{prompt_message}",
+        custom_map={"language": "Python"})
+    assert template.render({"language": "Rust", "prompt_message": "Mock Question"}
+                           ) == "Act as a Rust developer and respond to this question:\nMock Question"
+
+
+def test_render3():
+    template = PromptTemplate(
+        content="Act as a {language} developer and respond to this question:\n{prompt_message}",
+        custom_map={"language": "Python"})
+    with pytest.raises(MemorRenderError, match="Template and context are incompatible."):
+        template.render()